organizers list moved to the EventDetails page
