+++ app/src/main/res/layout/activity_event_details_page.xml
@@ -39,4 +39,21 @@
         android:layout_marginTop="5dp"
         />
 
+    <TextView
+        android:id="@+id/EditEventOrganizers"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:textAlignment="center"
+        android:textSize="20sp"
+        android:text="Organizers"
+        android:padding="10dp"
+        android:ems="10"
+        android:layout_below="@id/description"/>
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/organizers_recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_below="@id/EditEventOrganizers"/>
+
 </RelativeLayout>
