+++ app/src/main/res/layout/activity_event_edit.xml
@@ -139,22 +139,6 @@
                 </TableRow>
             </TableLayout>
 
-
-            <TextView
-                android:id="@+id/EditEventOrganizers"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="center"
-                android:textSize="20sp"
-                android:text="Organizers"
-                android:padding="10dp"
-                android:ems="10" />
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/organizers_recycler_view"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent" />
-
             <Button
                 android:id="@+id/EditEventSubmit"
                 android:layout_width="fill_parent"
